isContinuation needs a lot of work
We run into an issue where we can't read the restartrad.sph.orig file because it originates from a different simulation. The only way around that is to search in a different way.

For example, if a simulation with ncooling=0 is continued now with ncooling=3 then the output format changes, and we won't be able to read restartrad.sph.orig.