cleanup to file locking
Now any arbitrary number of processes should be able to lock a file any arbitrary number of times without stepping on top of other locks. This will create a lot of empty junk files, but those should both not be visible on the user's system by default and also not take up any space.