joining simulations is only allowed now if they share a restartrad.sph.orig file
This is to simplify the process by which joined simulations obtain their output files. Previously we tried sorting the output files by simulation times, but this proved to be far too expensive. Now we only check at what simulation time a simulation has started, and sort the joined simulations that way.