added a special minimum flux limit
ok, so if you do with "- minv 5" (i.e. minimum value of the flux for which you account to) and then do not account for all the rays which have flux 10 times less than 10^5 then luminosity is the same as with a higher range to the precision of the 4th digits but is obtained much faster. Again that is if the domain is limited to max coordinate in any direction being 2000 instead of

My condition that I imposed was

if z[i] > surf_d[ii][jj] and dist <= rloc[i] and tau[i] >= tau_skip and flux[i] > pow(10, min_value-1):
                        ray_id[ii][jj][ray_n[ii][jj]]=i
                        ray_n[ii][jj]=ray_n[ii][jj]+1