Reader update
The Reader class can now better detect the "write" statements in the StarSmasher Fortran source code that are responsible for producing binary file outputs ("out*.sph").

The method before was only barely sufficient and was not very robust. The new method introduces a FortranFile class that uses regular expressions to parse the code. All "write" statements are compared to the contents of the binary files to search for the most likely match.